Have the new way arraywidgets are delivered to the ArrayView setup and working. There's one last little warning I wannt get out of the way.
-Kurtis

+++ example/BasicExample/main.cpp
@@ -105,7 +105,7 @@ int main(int argc, char* argv[])
    * Here we create a Array object of 10 doubles
    * representing an initial guess for a linear solver.
    */
-  Array<int> doubleArray( 10, 0 );
+  Array<double> doubleArray( 10, 0.0 );
   My_List->set(
     "Initial Guess", 
     doubleArray,
